requirements.txt
livekit-agents
livekit-plugins-cartesia
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-beyond-presence
google-auth
google-api-python-client
python-dotenv
pytz